requirements.txt
-i https://pypi.org/simple
boto3==1.26.109
botocore==1.29.109 ; python_version >= '3.7'
cachetools==5.3.0 ; python_version ~= '3.7'
certifi==2022.12.7 ; python_version >= '3.6'
charset-normalizer==3.1.0 ; python_full_version >= '3.7.0'
google-api-core==2.11.0 ; python_version >= '3.7'
google-api-python-client==2.84.0
google-auth==2.17.2 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
googleapis-common-protos==1.59.0 ; python_version >= '3.7'
httplib2==0.22.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
idna==3.4 ; python_version >= '3.5'
jmespath==1.0.1 ; python_version >= '3.7'
oauthlib==3.2.2 ; python_version >= '3.6'
protobuf==4.22.1 ; python_version >= '3.7'
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9 ; python_version >= '3.1'
python-dateutil==2.8.2 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
requests==2.28.2 ; python_version >= '3.7' and python_version < '4'
requests-oauthlib==1.3.1 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
rsa==4.9 ; python_version >= '3.6'
s3transfer==0.6.0 ; python_version >= '3.7'
six==1.16.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
uritemplate==4.1.1 ; python_version >= '3.6'
urllib3==1.26.15 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
